refactor: set up logger and modularize code
