+++ .github/actions/deploy-to-s3/package.json
@@ -28,6 +28,7 @@
     "@types/node": "^14.14.31",
     "aws-sdk": "^2.848.0",
     "fast-glob": "^3.2.5",
-    "mime-types": "^2.1.29"
+    "mime-types": "^2.1.29",
+    "winston": "^3.3.3"
   }
 }
